feat(cli): port baml wrapper to clap and unify formatting via baml_term (B-886)
The baml wrapper hand-rolled arg parsing, help text, and error/warning
output, so its rendering diverged from baml-cli in the same invocation
stream. Extract the shared presentation pieces (CLAP_STYLING, ColorChoice
and agent-aware init_color, the error:/warning: printers) into a new tiny
baml_term crate that the wrapper, baml_exec, baml_cli, and the pack host
all consume, and port the wrapper's toolchain/self-update commands to clap.

The wrapper still dispatches on argv[1] so everything except its own
commands passes through to baml-cli byte-for-byte; clap only parses
wrapper-owned commands.

Deliberate behavior changes:
- wrapper usage errors follow clap conventions (styled error + usage,
  exit 2; bare 'baml toolchain' prints help, exit 2)
- wrapper main errors render via print_anyhow_error (bold-red error:
  header + cause chain) instead of anyhow's one-line alternate format
- the shared printers gate styling on stderr (for_stderr) instead of
  console's stdout default, matching where they write
- baml-cli's direct-invocation warning uses the styled printer
