+++ baml_language/crates/baml_cli/Cargo.toml
@@ -25,6 +25,7 @@ baml_lsp2_actions = { workspace = true }
 baml_lsp_server = { workspace = true }
 baml_project = { workspace = true }
 baml_release = { workspace = true }
+baml_term = { workspace = true }
 baml_type = { workspace = true }
 baml_version = { workspace = true }
 baml_workspace = { workspace = true }
