+++ baml_language/crates/baml_exec/Cargo.toml
@@ -18,13 +18,13 @@ doctest = false
 workspace = true
 
 [dependencies]
+baml_term = { workspace = true }
 baml_type = { workspace = true }
 bex_engine = { workspace = true }
 bex_vm_types = { workspace = true }
 anyhow = { workspace = true }
 borsh = { workspace = true }
 clap = { workspace = true }
-console = { workspace = true }
 indexmap = { workspace = true }
 serde_json = { workspace = true }
