+++ baml_language/crates/baml_term/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "baml_term"
+version = { workspace = true }
+publish = false
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+description = "Terminal output conventions shared by every BAML CLI surface"
+
+[lib]
+doctest = false
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+clap = { workspace = true, features = [ "color" ] }
+console = { workspace = true }
